Templates
Added templates for the display of data using bootstrap for styling and
display and jinja2 for displaying data passed to the template file.
Added an email.html file for users to input their personal message when
sending emails.
Added an invoice.html file to display a HTML representation of an invoice
which can then be convered into a PDF.
Added an invoice logging file to select data when logging a new invoice.
Added an item logging file to select data when logging a new item.
Added a invoices.html file to display an overview of each invoice and
provide access to links associated to each invoice.