Refractor Item
Refractored the Item class with restructured and documentation.
Added docstrings to each class and method in the Item class.
Renamed the item_code class variable to code for simplicity.
Separated SQL statements into temporary variables for a cleaner document.
Made the constructor raise a Key Error if an invalid item code is
provided.
Created the _generate_code method as a helper method for create.
Ensured that when an item is created it does not have an item code that is
already in use.
Following research, removed all standard getter methods in place of
properties where needed. Also updated associated files to reflect the
changes to the interface.