Refractor Person
Added docstrings to all classes and methods relating to the person class.
Ensured that a person exists before attempting to retrieve the person data
from the database, if a person does not exist in the database then it
raises a key error.
Added the create method to the Person class to allow easy insertion of new
people into the database.
Added the delete method to the Person class to allow easy deletion of
people from the database.
Following research, removed the getter methods and changed any variables
requiring getter and setter methods to properties.
Removed the __str__ method and started using the name variable to replace
this functionality.
Updated the template files to match the changes to this model as well as
previous model updates that were missed.