Pagination
To reduce the amount of database access required when loading the invoice
list pagination has been added to separate invoices into pages.

By default 5 invoices will load per page. This configuration will later be
moved to config.py.

To implement this change the url /invoices will act as a default page 1
and other pages are identified by /invoices/page/<pagenum>. The addition
of page is to prevent confusion with the similar /invoice/<invoicenum>.