Graph Modification and Email Fix
Modified the graph to plot money over the invoice number rather than the
date to give a better indication of change.

Added labels and styling to the graph and set it to draw as a xkcd figure.

Fixed email sending by updating the references to the person class with
the new interface.