Invoice Deletion
Upgraded the invoice classes delete function to include the deletion of
any associated items when an invoice is deleted. This will later be
changed to a database cascade system.

Added the URL /api/invoice/<invoiceID>/delete which when called will
delete the given invoice. This should be altered to use the DELETE method
in a more RESTFul design.