Item Log Upgrade
Upgraded the page and API for logging items.

To allow for a RESTFul upgrade to the item API a new method was added to
the item class, the dict method which returns a dictionary representation
of an item.

The /api/log/item url was moved to /api/item.

To create an item use a POST call to /api/item.
To delete an item use a DELETE call to /api/item.

The create and delete API methods return a json structure of the item.

The template for the item log page has been upgraded (this page will
eventually move to a model).

The template file has been updated with the JQuery library.

Logging an item is handled by javascript rather than a HTML form.

A new material design snackbar style element has been added using
bootstraps alert. When the ajax call to create a item is completed the
snackbar will popup alerting the user that it has been logged and provides
a button for undoing this action.

When the undo button is clicked the item will be deleted through another
ajax call after which another snackbar will pop up informing the user that
the item has been removed.