Improved RESTFul Item API
Improved on the RESTFul item API by updating response codes to match
standards, updating the delete URL to match standards and adding GET and
PUT methods for reading and updating items.

Altered the item api post call to return a 201 response to indicate that
the item was created.

Added a base get url for the item entity to return a list of all items.

Added a specific get url for retrieving information about each item.

Added a PUT url for updating the data of an item.

As a result of the new PUT url a update method was added to the item class
to allow variables to be updated.

Ensured that the GET, PUT and DELETE methods checked if an item exists
before performing an action, if an item does not exist it wil return a
404 response.

Also edited the template file to represent the new format of the DELETE
url.